Rename currentprocess to process
