+++ src/bin/rustup-init.rs
@@ -29,10 +29,10 @@ use rustup::cli::rustup_mode;
 #[cfg(windows)]
 use rustup::cli::self_update;
 use rustup::cli::setup_mode;
-use rustup::currentprocess::Process;
 use rustup::env_var::RUST_RECURSION_COUNT_MAX;
 use rustup::errors::RustupError;
 use rustup::is_proxyable_tools;
+use rustup::process::Process;
 use rustup::utils::utils;
 
 #[tokio::main]
