+++ src/lib.rs
@@ -72,7 +72,6 @@ fn component_for_bin(binary: &str) -> Option<&'static str> {
 pub mod cli;
 mod command;
 mod config;
-pub mod currentprocess;
 mod diskio;
 pub mod dist;
 pub mod env_var;
@@ -80,6 +79,7 @@ pub mod errors;
 mod fallback_settings;
 mod install;
 pub mod notifications;
+pub mod process;
 mod settings;
 #[cfg(feature = "test")]
 pub mod test;
